job #176 - Last minute fix for the feature to add built-in EEs.  The build server script did not like the fact that the folder had spaces in the name.
I was going to try to fix that script (create_bp_release.sh).  The
changes would be quoting some paths in the copy_included_files function.
But there is also a rather hair awk command in there that would have to
be reworked.  For the time being I opted to just take the spaces out of
the folder name here and will open an issue for the script fix.