#236 Added support for the concept of "slots" for ordered model elements
1) In create_metadata_elements_sorters.arc I generated 2 new functions:
public static Class<?> getOrderedElementTypeAtSlot(int slot)
public static int getOrderedSlot(Object element)

These work together to determine a unique "slot" number for each of the
ordered model element in the ooaofooa.  The model elements are
determined by using the T_TNS.SortingRelationNumber attribute.  Slots
are defined starting with a value of 1, and going from there.

2) In create_object_inspector.inc a new operation was generate into each
of the inspector classes.  The operation is:
 	public int getNextSlot();
This operation simply returns a count of the attrbutes+referentials for
the inspector class.  The count serves as the "next" available location
for elements added during merge.
 
3) Modified TreeDifferencer.java::getLocationOfElement() to remove the
previous code that was used as proof-of-concept, and replaced it with
calls to the new generated operations described above.