#259 We reviewed the implementation note
We reviewed this without the author, Keith.  We did record a major observation, but we will go ahead and promote this issue now because it is tied to #35 which is ready to promote and we want to get it in now.   We will leave this issue open until Keith returns and can address the observations.