Update Developer Getting Started Guide.md
job #46 I removed 2 places where the document referred to the mc repository.  This is not a public repository and should not be referred to in this document.  The mc source is currently found in the bposs repository.  In the fucture all source will move to the bridgepoint repository, but for now it is  in bposs.