job #37 The process_build.sh script performed 2 tasks: 1) called create_bp_release.sh to create the build 2) distribute and notify
I removed this script, and folded it's tasks into the main script,
run_build.sh.  I did this by simply calling process_build.sh instead of
the previous call to process_build.sh, and created a function in
run_build.sh named distribute_and_notify that performs the remaining
functionality.  The reason I did this is that the complexity in this
script (what little there is), is in the distribute_and_notify
functionality, so the change better encapsulates what is actually
happening and makes distribution and notification more clear.