Update headless_build.sh
job:#7613 The prior change introduced a 3rd build cycle to try and deal with dependency issues.  However, it simply touched the same file for the 2nd and 3rd builds, so those builds would be doing the exact same thing.  The result was that the problem we were trying to resolve (io.core and model.compare were not building) was not resoled.  I added a new function, build, that gets called each of the 3 times and we pass-in options for "clean" and "file_to_touch" to try and resolve this.