job:#7627 In the privious change, docgen.xsl was moved to the same folder as docbook.xsl so the docgen.xsl import specification is satisfied.  This doesn't really need to be done at runtime for each invocation, so instead this change has the packaging script move docgen.xsl to the docbook folder.
There was a large comment in the code
(Generator.java::runXsltproc())that describes why this was done like
this.   This change removed that comment, but I don't want it to be
lost, so I am placing it here:

// move docgen into the same folder as docbook.xsl.  This is done
because docgen.xsl includes docbook.xsl and 
// for that include to be resolved we have to have one of 3 things:
// 1) docgen.xsl and docbooks.csl are in the same folder (this is what
we have done)
// 2) Use the xsltproc --path option to add a path for docbook.xsl
//    Note: This was preferred, but failed because at runtime the --path
option was seen as an invalid
//          option.  this was odd because it works on the command-line,
but it does not work, at least in linux,
//          at runtime.
// 3) modify the include specific in doc.xml to have a full of relative
path to docbook.xsl.
//    Note: This was the option used prior to v5.0 of the tool.  Version
5.0 removed the post-processing of installation
//          that performed this modification, so this option is no
longer possible, and is was not a desirable option anyway.