job:#8238 I made several changes to further facilitate the ability to perform reconciliation of connectors.
I made a change to prevent the calling of operations needed only for
connector reconciliation during shape reconciliation.  This problem was
introduced with the new GD_ARS instances added to ui.glaphics/plugin.xml
explicitly for connector reconciliation.  To fix this I took advantage
of an attribute that was added specifically for connector reconciliation
named GD_ARS.elementExistsMethod. I renamed this attribute to
GD_ARS.connectorExistsMethod, and I modified the OAL to only attempt to
call this operations when that attribute is used. This prevents calling
it during shape reconciliation.

While making this change I did make a couple other small changes to
improve errors handling as well as to not call through the Cl_c.java
interfaces to the ooaofooa unnecessarily (when an operation is labeled
"not used" there is no reason to make a call, we know it will fail).