job:#8238 I made several changes
_- I fixed some problems that I made previously when I tried to isolate
the amount of GD_ARS data we use when reconciling graphics by using the
changed model element. The idea was valid, but the ooaofgraphics
instance selection that I made was not correct and this broke some of
the functionality that was working prior to me having made that
"optimization". 

_- I introduced GD_ARS.removeConnectors. This functionality was
previously inside GD_ARS.reconcileConnectorStep1, but it was complex and
self-contained enough to warrant this separate operation. This new
operation is still invoked in the same place, inside
GD_ARS.reconcileConnectorsStep1. The reason as we reconcile connectors
if the actual number of connectors encountered is 0 then we know we have
 a situation where we have a graphic element (GD_GE) that has no
associated ooaofooa model element and thus it needs to be removed.
Therefore this removed is done in the same pass over GD_ARS instances
with the first-pass of connector creation.

_- I removed the GD_ARS.reprocess attribute which was used to flag
conditions that require graphics reconciliation to be rerun in order to
account. The only current case for a "rerun" is where corrupt connectors
are encountered and removed. When this happens we do create an error log
entry.