job:#8238 Created a new entry point, GD_ARS.ReconcileAllGraphics, and restored the orginal GD_ARS.Reconcile to its orginal state.
This change is made so there are 2 entrypoints for graphics
reconciliation. One is the 
original that is used when a model change is made. The other is used to
reconcile all 
graphics in the model. This new entry point, ReconcileAllGraphics,
allows for graphics 
creation on the whole model without the need to have it invoked during
import when all 
model elements are being loaded.

1) I made a new copy of GD_ARS.Reconcile named
GD_ARS.ReconcileAllGraphics 
2) I modified GD_ARS.Reconcile to return it to its original
functionality. 
3) I modified ui/canvas/GraphicsReconcilerLauncher.java (the interace
into the 
ooaofgraphic's GD_ARS ) and made a copy of runReconcile(removeElements)
with an added
parameter runReconcile(removeElements, createAllGraphics). The orginal
now simply invokes
the new copy with createAllGraphics==false.
4) In GraphicsReconcilerLauncher.java::runReconciler(), when
"createAllGraphics" is true
we call GD_ARS.ReconcileAllGraphics. 
5) In GD_ARS.ReconcileAllGraphics we use the
CL::numElements/CL::GetElementID to call
S_SYS.{GetGraphicalRootCount()/GetGraphicalRoot()}to get the graphic
roots and 
use this to call GD_ARS.ReconcileCanvas with each one before we run the
rest of the 
graphics reconciliation.