job #8321 Changed cut/paste behavior to move.
I changed the behavior of cut/paste to make it move. To do this I
introduced a new abstract class, CutCopyPasteAction extends Action, and
I modified the  existing CutCopyAction.java and PasteAction.java classes
to extend this new class instead of extending Action directly. This
allowed me to move implementation from CutCopyAction "up" into this new
class to allow PasteAction to know when the user has selected cut vs
copy and  to performed the cut/paste in a single transaction that occurs
in PasteAction. I moved the code that was deleting elements out of
CutAction.java::postRun() and put this into Paste action at the point
AFTER the move has taken place. I added an attribute to
CutCopyPasteAction named ELEMENT_MOVE_SOURCE_SELECTION that holds the
IStructuredSelection made by the user during cut. This allows us to
deleted the selection without concern that the user may have pasted some
element under the same model root that it was cut from. It also removes
any other potential problems with lookup by ID that may have problems
because the element IDs are not changed.  In order to prevent the
element IDs from being changed I modified bp/io/CoreImport constructors
to pass a new parameter, boolean createIDsDuringModelImport. This flag
is used in the generated file ImportModelStream.java::finishLoad to
prevent the code that changes the IDs form running. There was a quite a
bit of fallout from this change simple because the change was made to an
archetype that generates a lot of different types of import classes.
Each place that instantiated one of these objects had to be modified to
account for the new parameter. Note that in almost all cases the value
of this is set to true, because we DO want to update the IDs in most
cases. The only other cases where we were not updating IDs was in model
merge.