job #8321  Don't allow move to the source PMC
1. Disable paste when the selected target is the same as the source.
1.1. bp/ui/explorer/actions/ui/Explorer{Cut | Copy |
Paste}Action.java::isEnabled() 
implements this behavior on behalf of the
org.eclipse.jface.action.Action abstract class.
This is where BridgePoint determines if the CME should be enabled or
not. Note that there 
is an analogous implementation for canvas in 
bp.ui.graphics.actions/Canvas{Cut | Copy |
Paste}Action.java::isEnabled()
1.2. The operation called out in #1 is implemented to check the source
and destination to
see if the CME should be enabled or not.
1.3. In the PasteAction case, the operations look to see if the
destination allows paste for 
elements in the source being pasted to the target. It does this by
calling an ooaofooa 
operation that is of the form {target model element
instance}.Paste{Source Model Element Name}
. The structure of this code was such that this behavior 
was essentially duplicated in {Explorer | Canvas}PasteAction.java. I
refactored this and 
"moved up" an operation named clipboardContainsPastableModelElements()
into the parent class
core/ui/PasteAction.java to facilitate adding the check to assure that
on paste of a cut the
source and target do not match.
2. In PasteAction.java::isEnabled() if a move is in progress do not
allow more than 1 
selected destination.