job #8321 The file move operation needs to happen after the memory model has been modified. We can not use cut/paste for move for multiple reasons.  Basically, it comes down to the fact that "delete + add != move".
1) In Datatypes/ModelEventNotification I added an new enum
DELTA_MODEL_ELEMENT_MOVE. Note that this enum is actually used in a
bitset implementation
so the value of this new element is 16384 ( 0b0010000000000000). I then
introduced the 
Java code required to use this new ModelDelta type. 
2) I introduced class ModelElementMovedModelDelta extends
BaseModelDelta, and added a new interface operation, void
modelElementMoved(ModelChangedEvent event, IModelDelta delta), to
IModelChangeListener.  
3) I updated places as needed that had to implement this interface. The
only one that has a non-empty implementation is
ComponentTransactionListener.java, and in it is where I added code that
performs the folder move when the transaction ends. 
4) I introduced
ComponentTransactionListener::movePMC(NonRootModelElement elementMoved,
NonRootModelElement destination) that performs this task.
5) I introduced ModelRoot.java::fireElementMoved() that is used to
signal listeners of the 
ModelElementMovedEvent
6) I modified PasteAction.java::run() to signal the TransactionManager
with the ModelElementMovedEvent after the move has been performed in the
loaded model 
7) I introduced NonRootModelElement.java::getPE() to get the
PackageableElement instance 
associated with the NonRootModelElement instance (if there is one, null
if not).  Putting
this operation here allows us to use the instance list to find the PE_PE
instead of having to do it via reflection or some other means that would
have to account for the specific type of the NRME. I modified
PasteAction to take advantage of this new operation.