job #8470 In GD_ARS::ReconcileAllGraphics we have a 3rd pass that currently is only used to create the linked class association. The reason is that the association must be created before the linked class association can be.  Of course subtype is similar, it can not be created until the supertype association exists. I therefore deferred the subtype creation to pass 3 as we do linked association.
In GD_ARS.reconcileConnectorsNoExistingGraphics there is a special-case
for the linkedAssociation case (the case where a connector is connected
to a shape on one side but a connector, not a shape, on the other). This
Subtype situation had to be added to this special case. 

after subtype connector creation the represents for the new connector
created (GD_GE) is not being set properly, it is null. The reason seems
to be that the call to CL::getInstanceFromOOAId for this connector type
(subtype) is failing to find the ooaofooa instance.

The reason is that {O_OBJ | O_IOBJ}.getSubtypeid() was returning the
Rel_ID. This does not uniquely identify the subtype instance. The
classaslink was used as a guide, but it's cardinality is 1 on both
sides, so it can use the Rel_ID, but subtype needs the OIR_ID. The next
problem is that now that we use the Subtype's OIR_ID we DO find the
ooaofooa instance, however, now when we go to find the "other side"
there is no match because the superype's id is using the rel_ID.