Make VerifierInvocationHandler call the setters when marshalling out structured data type members
When performing an audit the Auditor will complain if no getter or setter is present, but only the getters were implemented in the invocation handler.
This will make the marshalling out work like the marshalling in, if a setter is present, try to use it. If that fails, fall back to writing the field directly.