job #8321 I updated the move pseudo code
I made it a little more high level, and removed the parse after move. That was not a requirement and during discussion of the parse all after move, we determined it was too time-consuming to force it. The user may be performing several move operations, they can choose when to parse.