job #8321 I put back the code to update proxies during model element move.
The decision has been made to update the proxies paths during move
instead of leaving the stale. The advantage of leaving them stale was
that RGOs would not be checked-out and dirtied just to update proxy
path, but the disadvantage was that leaving stale proxy paths seemed bad
architecturally because proxies are still being used in BridgePoint, so
it was decided they should be left in a consistent state. A follow up
project may remove proxies altogether (#8443 #8729), but it is a
stand-alone project.