job #8321 I uniquely named the Model Element Move transaction to distinguish it from the Paste transaction to allow the transaction to be filtered from graphics reconcilation.
We do not need to run graphics reconcilation on a Move.
Note that while considering removing move from reconciliation, the
question came up about 
disabling reconciliation on copy/paste. It is worth noting that in the
case where elements are copied from ME, all graphics ARE also copied
(even though the selection was in ME and not canvas), so it is not the
reconciler that "recreates" graphics during copy/paste it
is the copy to the clipboard (used by copy/paste) that is smart enough
to get any graphics associated with the ME selection. However, on paste
in that situation the container symbol for the element being pasted IS
created by graphics reconciliation. Therefore, for now only move is
filtered.