job #8321 I fixed a problem that was preventing RGOs from being updated in the case where a selected is an RTO that is not a model root. To resolve this there were 2 changes:
1. Error in TransactionListener::movePMC the code was not properly
setting 
the PMC resource file in the case where the source was not a root model
component.
2. Change in TransactionListener::persistRenamedME() that improves how
we 
look for RGOs. While making this change I reused some code from the
TransactionManager
that does a better job of collecting the PMC children then the code here
was doing, plus 
there was no reason to duplicate the functionality.