job #8321 I introduced a new operation, NonRootModelElement.java::move_unchecked(ModelRoot destination). This operation is similar to the delete_unchecked() operation in that it operates on the InstanceList associated with the NonRootModelElement instance. The move_unchecked() operation removes the NRME instance from it's instance list and adds it to the InstanceList of the given ModelRoot.
I modified TransactionListener.java::movePMC to use the new
move_unchecked() operation
in the situation where we are moving a NRME that is not a ModelRoot and
whose rto is a
DataType_c instance. In this special case, There are actually 2 NRME
instances that must 
be moved. The instances are : DataType_c, and <specialized user type>
(an example of the
specialized type is UserDataType_c).

Additionally, I modified movePMC() so the the in-memory portion of the
modifications come
before the actual folder move (when there is a folder move). I did this
to facilite the
future step where we will enable undo for this move operation. I added a
big comment
at that point to describe what needs to be done to implement the undo.

Finally, I made a change in the folder move portion of movePMC() to
differentiate the 
move of an element that is a model root from an element that is not a
model root.
1. Moved Element is a Model Root
When the element is a model root, the code updates the moved NRME's
resource file 
(within its PMC) to point to the new folder. 

2. Moved Element is NOT a model root
When the moved element is NOT a model root the moved element is being
moved to a different
PMC, so the NRME's PME is switched to be the PMC of the destination
element.