job #8321 I introduced an enum, BPTreeType.
This new enum is used to represent the type summaries that are shown for
each model root. For example:
  Ooaofooa Model Elements
  Graphical Model Elements
  Parser Instances
  Runtime Instances
  
Prior to this change, these were static strings help inside the
BPProjectTree class. I made the an enum instead of a String to make it
clear what is being passed around and returned at the spots where this
is used. There is currently a bug that is causing each of these summary
lines to contain an accumulation of not just what is in the "current
root", but what is in all roots that were listed "before" the current
one. This small change is made to help clarify the code in an effort to
fix this problem.