job #8321 I made a couple of changes based on review of the BridgePoint architecture.
The architecture review information has been captured in the FAQ under a
new section, BridgePoint Architecture. The changes made here were in the
ComponentTransactionListener and in the CancePasteAction. Both changes
were specific to MEM. The changes were made because the picture in my
mind of where the InstanceList associated with a NRME was not correct.
However, the net effect of this change on the issue at hand was nil. The
reason the net effect was nil is that the code uses nrme.getModelRoot()
to obtain the model root therefore even though the picture in my mind
was not correct, the ModelRoot being obtained was the correct root each
time. This change is still very useful because it does clarify the code.
However, at this time, the empty white box after moving a NRME back and
forth a couple times is still a problem. The issue, at this point does
not appear to be the changes in place. The issue seems to be the the
ModelChangeListener is running after the transaction completes, and it
is creating an additionally GD_GE instance for the moved element. That
seems to be the issue. I will also note that I modified the Instance
Population Viewer so that it now shows graphical instances. This tool
does allow the extra GD_GE instance to be seen easily (debugger no
longer needed). This should aide in tracking down this problem.