job #8449 I made a change to assure that we treat source and destination sides of a connector properly in cases where connectors are directional.
To do this, I created a new realized bridge, 
boolean isDestination(Instance shapeInstance, Instance
connectorInstance). 

This bridge is implemented in Cl_c.java. Each of the 2 instances passed
into this routine are ooaofooa NonRootModelElement instances. This
routine returns false by default. It returns true only in the case where
the shape passed is in a StateMachineState, the connector
is a Transition, and the state given is the destination state for the
given Transition. It returns false for all other cases.

This bridge is used in GD_ARS:reconcileConnectorsNoExitingGraphics()
after we have all the information needed to create the connector, but
just before we actually make the call to create the new connector. This
bridge determines if "side1" of the connector is the really the
destination instead of the source for the connector. If 
it is the destination, then we swap side1 and side2 just
before we create the connector so that we get the source and destination
of the connector correct during creation. Note that at least currently,
it is only transitions that are 
directional, so it is the specific case transitions where the new
isDestination() bridge may return true, in all other cases it returns
false and therefore no swap occurs.