job #8484 I tested and fixed 2 problem in the new autoCreateSatisfactions functions. 1. When we create a satisfaction, the satisfaction is persisted in the same package that holds the diagram where the satisfaction takes place. In the case of autoCreateSatisfactions, this is the package that is passed-in to the routine.
2. As we auto-create satisfactions a change was made to assure that once
a satisfaction is created for one side of the satisfaction we do not
create a second one for the other side. This fixes a bug that was
causing 2 C_SF instances for each provided/required pair instead of 1.

It is worth noting that this operation is specific to MASL import. In
MASL there will only ever be 1 satisfaction for a given provision.