job #8955 I removed some obsolete model updgrade code that had an old TODO comment with it. It looked lie this:     if (table.equalsIgnoreCase("C_C") && !actualVersion.equals(persistenceVersion) && (!actualVersion.equals(bp7_1_5VersionNumber)) && !actualVersion.equals(bp7_1_4VersionNumber)) { //$$NON-NLS-1$$       // TODO - This code was added to upgrade C_C instances, but only for       // these instances when created before 1.5 ...
That code was destined to break the product the next time someone bumped
the persistence version, and sure enough it did so. That code, as the
TODO says, was used to update mode with file persistence older than 1.5,
and BridgePoint has forced user upgrade of model since then, thus the
code is obsolete and is being removed.