I made 2 changes in ImportHelper.java::resolveMASLproject() to be more graceful about error conditions during mask import. 1. We now use a single transaction for the masl import (previously there were multiple), and if anything goes wrong during the import we abort the transaction and inform the user.
2. If a package is marked as masl_project, and we do not find a matching
domain for a 
component reference in the package then we inform the user with a
message to the console. 
This is not fatal, it is just a warning to the user. Here is the text of
the message:
String msg = "The following Component Reference was not assigned: " +
cl_ic_name;
       msg += "\nThe domain was not found in the workspace.";
       msg += "\nAssure that Interproject References are enabled and
that the domain has been imported.";