job #8978 I modified bp.ui.graphics/plugin.xml and added GetSatisfactionID in Requirement and ImportedRequirement. In the ooaofooa I implemented these 2 operations. While editing the ooaofooa I saw that getSatisfactionCount was implemented in Provision, and it is not being used there, so I removed it.
In canvas/ooaofgraphics I added support for satisfactions in
GD_ARS.ReconcileAllGraphics by assuring that the satisfaction
reconciliation routines only get used in the 3rd pass (we have to assure
the other instances have been created before satisfaction). In
GD_ARS.reconcileConnectorsNoExistingGraphics I made similar
modifications to assure we handle satisfaction reconciliation in pass 3.
I introduced CL::getProvisionFromSatisfaction to find the Provision
associated with the satisfaction. At the point where the Provision is
obtained we have all the information need to reconcile:
1. GE_GE associated with the Provision
2. GD_GE associated with the Requirement
3. The OOA Satisfaction_c instance
We those 3 pieces of information we can use GD_MD.newConector to create
the GD_AOS for the satisfaction and move the requirement to the
provision. The GD_AOS links the ball and cup, but we have to move the
requirement so they appear together. I left the modification of
GD_MD.newConnector for the next step.