job #9354 Added no-arg constructor
The work done for issue #9316 via pull request
https://github.com/xtuml/bridgepoint/pull/388 introduced a new
constructor for class PullSynchronizationChanges.java (the
implementation for the Sync With Library CME). As a CME, a no-arg
constructor is needed for this too. I introduced one in this change.
This clears up an NPE that is observed when right-clicking a model in
ME. This NPE is caused when the Eclipse infrastructure attempts to see
if this CME action should be displayed or not.