job #9354 Ignore problem markers associated with multi-buffer editor
The work done for issue #9191 view pull request
https://github.com/xtuml/bridgepoint/pull/375 introduced changes to
allow problem markers in the problems view to open the single-buffer
masl editor instead of opening the masl file itself. This change causes
NPEs to occur any time the masl error finds errors in a ".mod" file. To
resolve this, this change causes only error markers for ".masl" and
".xtuml" to be used to associate masl errors with model elements in the
ooaofooa. 

As part of this change, the file resource that the error is actually
coming from is now being passed along into the routine that links the
error to a ooaofooa resource.