job #9354 Always create an error marker when xtext tells us
If we can't find the xtuml resource then use the default marker creation

The is a change associated with #9191 pull request
https://github.com/xtuml/bridgepoint/pull/375 . The work done in #9191
was only considering error found in .masl files where the .masl file 
was associated with an error in an xtuml file that was in the same
folder. This assumption is bad now that masl projects have been
introduced.  This change improves error marker behavior now that masl
project support has been introduced.