job #9354 In the section that the calling masl rename/refactor I made a few changes: 1. I removed the thread from thsi code. The fact that there needs to be a thread, and there does, is not a details that belongs in this class. I moved this thread to RenameParticipantUtil.java class where such details of masl refactor/rename are more appropriate.
2. We have a check to see if it is necessary to call masl refactoring or
not. This check was duplicating some complex, hand-craft code. I
refactored to removed this duplicate code.

3. This code explicitly ENABLED the BridgePoint resource change
listeners which reload model files based on change notifications. I
reversed this to assure that the resource change listeners are DISABLED
and I and a comment why it was done. The reason is that masl
refactor/rename does make changes to files, and the resource change
listener was reloading the file BEFORE this BridgePoint transaction
persisted the in-memory change. This causes the in-memory change to be
wiped-out before the call to persist(target), which was the root cause
of masl files being updated and xtuml files NOT being updated (example:
class model file when attribute is changed). This is a race condition
betweeen the reload and the persist(target) call.