job #8442 - Stop generating the context menu schema.  The following note is added to generate.xml:
The python-based generator updated in May 2017 started performing batch
relate which caused some issues with the PEI CME data which we create  	
and directly set referential attributes in.  In the Context Menu model 	
there are a number of relationships that are formalized but never   	
traversed.  To get around the problem of the batch relate overwritting 	
our pre-populated referentials we no longer want to specify these   	
relationships in the schema, but schema gen does not know how to ignore	
formalized relationships, so we are choosing to not recreate the schema
on every build. Instead, we are committing the schema to revision
control, which means any changes to the context menu model will
necessitate hand-edits to this schema. See issues 8442 and 9481.