job #8442 - A while back Bob started some work in the io.mdl generation to stop using persisted gen db file and multiple build targets and instead just generate all the code in one target without a persisted gen db.
That work was never fully implemented.  This change completes that work
and puts it into active use.  This was done because at the time of this
change the pyrsl generator was not persisting the gen db like we needed,
and this works around the problem.  The change is not fully complete
because the generated code is not quite right yet.