job:#9556 Add support to control importing into a workspace
I added a couple of properties to an existing parameters structure that
is used to manage how we import a project. These new properties will be
used to control importing into the workspace and launching the help
after import. Note that control of launching help after import is not
required to test this issue, but I observed that we are always launching
help during test even when we don't need to, and this causes extra test
time and work that is not needed. Note that I did leave this behavior on
by default, so it does not effect product functionality.