job: 9749 Introduced review observations
Note that we used a google doc to review and we worked together to raise the observations in this google doc. The format of this document is not "normal" for our process, but this was a very large document with a 2-hour review so we choose to do the review in this way.