job #9919
Removed "Set Model Compiler" out of the BridgePoint Utilities menu. It is still placed with the org.xtuml.bp.ui.context-internal group, just not under the BP menu. The plugin does not seem have visibility to the BridgePoint Utility menu for contributions that are not part of Model Explorer and this change resolves that problem.