job:#9958 switch "succeeds" to "precedes"
In Interface::publishOoperationFromFunctoin in the section that gets the
last interface operation that was just created the OAL was using
"succeeds" and should I been "precedes". When an existing interface was
updated in the situation where the function has multiple parameters not
all the parameters were being published corrected. This change resolves
this.