job:#10161 Fixed a race-condition deadlock
This is simply a save so I do not lose this .java change (though it is a
generated file).

I introduced an override for executeOperation() and deferred calling
BridgePoint Save until the super.executeOperation() has completed. The
reason is that when this operation begins, no locks are held by this
thread, but super().executeOperation() aquires a workspace lock and then
calls doSaveDocument() in this class to save the file. doSaveDocument is
where code existed to call into BridgePoint to save the document (it did
this with an asyncExec()). However, BridgePoint also aquires a workspace
lock in a different thread. There was a race condition here where if the
thread running MASLSnippetEditor::doSave() did not exit before the
BridgePoint thread attempted to aquire the workspace lock then the
BridgePoint thread (BridgePointTransactionEndedJob) was blocked waiting
for the workspace lock that was held by the MASLSnippetEditor thread.
Furthermore, when the deadlock did NOT occur, behavior was not as
expected because the MASLSnippetEditor::doSave() (inherited from
XtextEditor.class), after calling super.doSave() calls
calback.afterSave() to notify resource listeners that the editor has
been saved, and in this senario that worked, the save was actually NOT
complete, so though no deadlock occured here (at least in this
reproducible case, potential for other problems exists here.