job:#4896 Introduced the path_spec rule
During design discussions it has been determined that the grammar called
out in this issue's analysis as a suggestion is not the best path
forward. The user should not be forced to enter a fully-qualified path
every time they must disambiguate an element, they simply should be
required to enter enough information to disambiguate conflicts.
Additionally, from a path point of view, the grammar should not need to
know the difference between a model (S_SYS) and a package, because both
are simply part of the path (and perhaps eventually S_SYS may go away).
With these things in mind, the grammar has been updated.  Note that as
part of this change, the predicate defined in the rval rule for
scoped_data_type has been changed to be scoped_access.