job:#4896 Decision made for full-path specification baby-step
I removed the "!" from rule scoped_path.

This change puts the scoped_path rule in the AST allowing us to be more
specific in the problem marker about where an error happened. With this
change, there was fallout-out in the OAL Validation Functions that was
handled. 

Additionally, at this point a decision has been made to change the OAL
such that a System name must be specified as the first element of a
path_specification.  We made this decision because, while the grammar is
in place that supports the (more-flexible) approach that would allow the
user to specify only as much path as is necessary to disambiguate a
duplicate, the latter approach is not a requirement os this work, and it
introduces risk to the project. Where we are currently is a very nice
and clean "baby-step". We may indeed circle-back and change this, but in
the spirit of the "baby step" approach (and a solid deliverable) we are
moving forward with this much right now.