job:#4896 Introduced documentation from old issue branch
The analysis of this issue was originally done as a "dnt" because, at
the time, it was thought that the proposed "simple solution" was going
to be the one taken. Review of that not resulted in a different
decision, the decision to solve the problem by allowing the user to
specify a path, in OAL, to disambiguate a duplicate. When this happened,
an update to the note was made and that update was put in the form an an
"ant" because the work was larger and more analysis was required. This
analysis was completed and successfully reviewed. With this commit I am
simply moving the documentation from the stale branch to the new branch
where design and implementation has occurred.  BTW, the review of this
analysis happened on Feb 21, 2018. Here is a link to the old commit of
the review observations:
https://github.com/rmulvey/bridgepoint/commits/4896_scoped_enum/doc-bridgepoint/review-minutes/4896_scoped_constants_ant.rvm.md