job:#10529 Watch for RuntimeException in adoptSelection
The Eclipse AdapterManger has added some assertions to it's runtime
checking in its AdapterManager implementation. This is causing unit test
failures in teardown() in places where this is now being thrown. I added
catch's to this to prevent this.  Note that it is Model_c that it is
happening for now (this did not happen in prior versions of Eclipse).