job:#10345 Correct problem LIKELY leading to Widget is disposed errors
The BridgePoint tool has suffered from sporadic "Widget is disposed"
errors for a long time. These errors are not fatal and therefore they
are often ignored. In looking at a specific instance of a widget is
disposed for a test problem I discovered a good article:
https://stackoverflow.com/questions/27289133/how-to-solve-widget-disposed-exception
. This article described that any time a selection listener is used by
creating a class that "extends ViewPart" one needs to assure that in the
overridden "dispose()" operation the selection listener(s) are disposed.
While the BridgePoint tool was doing this, it was calling the
super.dipose() first in these dispose() operations and then destroying
the created listeners after. This is of course very dangerous as the
super.dispose() may very well have done some cleanup. In these cases, I
moved super.dispose() to be the last call instead of the first call.