Support Dark Mode
Current implementation looks at the current theme and
applies default colors accordingly.

NOTE: These changes do not listen for theme changes, but
we already get a notification from eclipse to restart before
changes take full effect.  Later support could be extended
to adjust when the theme changes.

Old workspaces will consider the previous defaults as user
configured due to the ui.text design.  New workspaces can be
created or in the OAL editor preferences "Restore Defaults"
can be used.

Graphics also uses a new default color scheme when dark mode
is used.

Support projects named differently than actual folder structure.

Eclipse allows you to name a project independently of the folder
on disk.  When you have this situattion Import Model Component
will create unresolvable proxies.  This can create an issue with
many areas of the tool, but the one being fixed here is that global
types are not available as the results are added to the proxy and
not the real element.  There is only one line that needs changing
to address this issue and in that case we were taking an eclipse
resource an converting it to a filesystem object.  This was unneccesary
as the full path of the resource object is the exact path we need.